FAQ: Fixed broken sw link and minor typo.
The link to the StarlingX Software under the "Where is the
Code" heading of the FAQ was broken (404).  I added the
correct URL.

Additionally, I spotted a typo in the following section
"Can I Contribute to It? How?" heading.  I fixed that
typo by adding a space and making the sentence end
cleaner.

Closes-Bug: #1808431

Signed-off-by: Scott Rifenbark <srifenbark@gmail.com>